% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{asesmso}
\alias{asesmso}
\alias{asesmso-package}
\title{asesmso: An R Analytics Toolkit on MS Office for ASES}
\description{
An analytics toolkit comprising of a series of work functions on MS Office operations.
The toolkit has a series of featured funcionalities specifically designed for ASES.
}
\details{
This package is comprised of \cr \describe{
  \item{MSO Enums}{\code{\link{enum}()}}
  \item{PPT}{\code{\link{getPptElems}()}}
  \item{MSO Charts}{\code{\link{getMsoChartDims}()}, \code{\link{getMsoChartSeries}()},
    \code{\link{getMsoChartSourceData}()}, \code{\link{getMsoChartType}()},
    \code{\link{setMsoChartScale}()}, \code{\link{setMsoChartSeries}()},
    \code{\link{refresh_ppt_charts}()}}
  \item{Scales}{\code{\link{optimizeScaleRange}()}, \code{\link{optimizeScaleUnit}()}}
}
}
\seealso{
\pkg{\link{aseskit}}
}
\author{
\strong{Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
