% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_matrix.R
\name{getCoreDataMtx}
\alias{getCoreDataMtx}
\alias{get_core_matrix}
\title{Shrink the matrix to its core area}
\usage{
getCoreDataMtx(mtx, gapWidth = 2, emptyString = c(NA, NaN))

get_core_matrix(mtx, gapWidth = 2, emptyString = c(NA, NaN))
}
\arguments{
\item{mtx}{the raw matrix to process.}

\item{gapWidth}{numeric, width of the "gap" rows or columns. Default 2.}

\item{emptyString}{character vector, how to define the "gap".
Default \code{c(NA, NaN)}}
}
\value{
A shrinked matrix
}
\description{
A matrix may contain separate data areas. This function will fetch the upper
left one as the core area, retaining the original row index and column index.
It is usually used after \code{\link{getMsoChartSourceData}}.
}
\examples{
\dontrun{
## The source matrix
m <- matrix(c(NA, NA, NA, NA, NA, NA,
              NA, 0,  0,  NA, NA, 0,
              NA, 1,  2,  NA, NA, NA,
              NA, NA, NA, NA, NA, NA), byrow=TRUE, ncol=6)

## There are two areas, with a gap of 2 columns
getCoreDataMtx(m, gapWidth=2)  # yields the left upper 2*2 matrix
   2  3
2  0  0
3  1  2

## If the gapWidth is larger than 2, then the gap will be included in
## the data area
getCoreDataMtx(m, gapWidth=3)  # yields:
   2  3  4  5  6
2  0  0 NA NA  0
3  1  2 NA NA NA

## If we include 0 in emptyString, then
getCoreDataMtx(m, gapWidth=2, emptyString=c(NA, NaN, 0))  # yields
   2  3
3  1  2
}
}
\seealso{
\code{\link{getMsoChartSourceData}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
