% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{getMsoChartSeries}
\alias{getMsoChartSeries}
\alias{get_msochart_series}
\title{Get the series parameters from a msoChart object}
\usage{
getMsoChartSeries(oChart, format = c("list"))

get_msochart_series(oChart, format = c("list"))
}
\arguments{
\item{oChart}{msoChart object (COMIDispath class)}

\item{format}{(deprecated) "list"}
}
\value{
A list comprising of lists of name (Name), label (XValues),
value (Values) and/or weight (BubbleSizes) and orient (series on rows or columns),
e.g.: \cr
$`1``\cr
\describe{
\item{$ name}{\tabular{lllllll}{
 \tab "ul1" \tab "ul2" \tab "lr1" \tab "lr2" \tab "sheet" \tab "formula" \cr
[1,] \tab "1" \tab "2" \tab "1" \tab "2" \tab "Sheet1" \tab "Sheet1!$B$1"}
}
\item{$ label}{\tabular{lllllll}{
 \tab "ul1" \tab "ul2" \tab "lr1" \tab "lr2" \tab "sheet" \tab "formula" \cr
[1,] \tab "2" \tab "1" \tab "3" \tab "1" \tab "Sheet1" \tab "Sheet1!$A$2:$A$3" \cr
[2,] \tab "5" \tab "1" \tab "5" \tab "1" \tab "Sheet1" \tab "Sheet1!$A$5" }
}
\item{$ value}{\tabular{lllllll}{
 \tab "ul1" \tab "ul2" \tab "lr1" \tab "lr2" \tab "sheet" \tab "formula" \cr
[1,] \tab "2" \tab "2" \tab "3" \tab "2" \tab "Sheet1" \tab "Sheet1!$B$2:$B$3" \cr
[2,] \tab "5" \tab "2" \tab "5" \tab "2" \tab "Sheet1" \tab "Sheet1!$B$5" }
}
\item{$ orient}{[1] "vertical"}}
}
\description{
Each series in a msoChart object can be represented as a formula:
"=SERIES(Sheet1!$B$1,Sheet1!$A$2:$A$3,Sheet1!$B$2:$B$3,1)", comprising of
Name, XValues, Values (and/or BubbleSizes) and Index. This function can
retrieve the series parameters in list form.
}
\note{
This function will call MS Excel process to communicate with the mso
file via COM interface silently.
}
\examples{
\dontrun{
ppt <- RDCOMClient::COMCreate("Powerpoint.Application")
pres <- ppt$Presentations()$Open(<some ppt file>)
slide <- pres$Slides(2)  # the 2nd slide
shape <- slide$Shapes(3)  # the 3rd shape
if (shape[['HasChart']] == -1) {  # if the shape contains msoChart
    chart <- shape[['Chart']]
    getMsoChartSeries(chart)
}
}
}
\references{
Refer to \code{\link{cell_limits}} in \code{\link{cellranger}} to understand
the data structure ul1, ul2, lr1, lr2, sheet, ...
}
\seealso{
\code{\link{RDCOMClient}}  \code{\link{cellranger}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
