% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_source_data.R
\name{getMsoChartSourceData}
\alias{getMsoChartSourceData}
\alias{get_msochart_sourcedata}
\title{Retrieve source data matrix from the msoChart object}
\usage{
getMsoChartSourceData(oChart, retainCoreAreaOnly = FALSE, gapWidth = 2,
  emptyString = c(NA, NaN), ...)

get_msochart_sourcedata(oChart, retainCoreAreaOnly = FALSE,
  gapWidth = 2, emptyString = c(NA, NaN), ...)
}
\arguments{
\item{oChart}{msoChart object (COMIDispatch class)}

\item{retainCoreAreaOnly}{Logical, only obtain the upper left core area of the
matrix. Default FALSE. If TRUE, the function \code{\link{getCoreDataMtx}} will
be called with preset arguments (gapWidth=2, emptyString=c(NA, NaN)).}

\item{gapWidth}{numeric, width of the "gap" rows or columns. Default 2. If
\code{retainCoreAreaOnly} = TRUE, then take \code{gapWidth} into effect.}

\item{emptyString}{character vector, how to define the "gap".
Default \code{c(NA, NaN)}. If \code{retainCoreAreaOnly} = TRUE, then take
\code{emptyString} into effect.}

\item{...}{Other arguments.}
}
\value{
A character matrix with character rownames and colnames.
Empty cells are converted to NA.
}
\description{
The function calls \code{\link{RDCOMClient}} to retrieve the source data from
the embeded Worksheet of a msoChart object in the form of character matrix. It
basically creates COM objects step by step and fetch \code{Chart[["ChartData"]][[
"Workbook"]]$Worksheets(1)[["UsedRange"]][["Value"]]}.
You may then process the raw matrix using \code{\link{getCoreDataMtx}}.
}
\note{
\itemize{
\item Currently \code{\link{RDCOMClient}} cannot retrieve non-ASCII characters.
\item By default, it only fetches the whole \emph{UsedRange} of the source data
 from the \strong{1st} worksheet.
\item It will open an MS Excel process to communicate via COM interface silently.
}
}
\examples{
\dontrun{
ppt <- RDCOMClient::COMCreate("Powerpoint.Application")
pres <- ppt$Presentations()$Open(<some ppt file>)
slide <- pres$Slides(2)  # the 2nd slide
shape <- slide$Shapes(3)  # the 3rd shape
if (shape[['HasChart']] == -1) {  # if the shape contains msoChart
    chart <- shape[['Chart']]
    getMsoChartSourceData(chart)
}
}
}
\seealso{
\code{\link{getCoreDataMtx}}  \code{\link{getCoreDataMtx}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
