% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chart_type.R
\name{getMsoChartType}
\alias{getMsoChartType}
\alias{get_msochart_type}
\title{Get the ChartType of a msoChart object}
\usage{
getMsoChartType(oChart)

get_msochart_type(oChart)
}
\arguments{
\item{oChart}{msoChart object (COMIDispatch class)}
}
\value{
If the ChartType is valid, then a named numeric is returned. If no
valid ChartType can be found, returns \code{NULL}.
}
\description{
If the msoChart is a pie chart, then this function will tell its formal chart
type: \cr
xlPie \cr
5
}
\examples{
\dontrun{
ppt <- RDCOMClient::COMCreate("Powerpoint.Application")
pres <- ppt$Presentations()$Open(<some ppt file>)
slide <- pres$Slides(2)  # the 2nd slide
shape <- slide$Shapes(3)  # the 3rd shape
if (shape[['HasChart']] == -1) {  # if the shape contains msoChart
    chart <- shape[['Chart']]
    getMsoChartType(chart)
}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
