% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_range.R
\name{optimizeScaleRange}
\alias{optimizeScaleRange}
\alias{optimize_scale_range}
\title{Auto suggest a range to beautify the axes of a masChart object}
\usage{
optimizeScaleRange(x)

optimize_scale_range(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector \code{c(min, max)}
}
\description{
Given a vector, automatically define the range to beautify the axes scale.
You can then use \code{\link{getOptScaleUnit}} to get the optimized major unit
of the scale.
}
\examples{
\dontrun{
optimizeScaleRange(c(3.1, 5.6))    # c(0, 6)
optimizeScaleRange(c(-4, -2))      # c(-4, 0)
optimizeScaleRange(c(-4.1, 15.6))  # c(-5, 20)
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
