% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_unit.R
\name{optimizeScaleUnit}
\alias{optimizeScaleUnit}
\alias{optimize_scale_unit}
\title{Auto suggest a major unit number to beautify the axes of a masChart object}
\usage{
optimizeScaleUnit(...)

optimize_scale_unit(...)
}
\arguments{
\item{...}{you can either pass \code{c(min, max)} or \code{min}, \code{max}}
}
\value{
a numeric scalar
}
\description{
Given a numeric range, automatically define a number that can serve as the
major unit to beautify the axes scale.
You are recommended to use \code{\link{getOptScaleRange}} beforehand to get
the optimized range of the scale.
}
\examples{
\dontrun{
# Both of the following expressions are OK
optimizeScaleUnit(0, 30) # 10
## or
vec <- c(0, 30)
optimizeScaleUnit(vec)   # 10
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
