% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_ppt.R
\name{refreshPptCharts}
\alias{refreshPptCharts}
\alias{refresh_ppt_charts}
\title{Automatically refresh msoChart in a PPT presentation}
\usage{
refreshPptCharts(files = NULL, orient = c("vertical", "horizontal"),
  autoFill = FALSE, save = TRUE, ...)

refresh_ppt_charts(files = NULL, orient = c("vertical", "horizontal"),
  autoFill = FALSE, save = TRUE, ...)
}
\arguments{
\item{files}{character vector of ppt files. If NULL, a GUI wizard will launch.}

\item{orient}{character, "vertical" or "horizontal". How the series is generated.}

\item{autoFill}{logical, whether fill the series with the entire row/col. Default
FALSE}

\item{save}{Logical, whether save the file after processing. Default TRUE.}

\item{...}{other arguments, e.g., \code{gapWidth} and \code{emptyString} for
\code{\link{getMsoChartSourceData}}.}
}
\value{
Nothing
}
\description{
It calls \code{\link{setMsoChartSeries}} and \code{\link{setMsoChartScale}} to
match source data with chart series and axes scales.
}
\examples{
\dontrun{
ppt <- RDCOMClient::COMCreate("Powerpoint.Application")
pres <- ppt$Presentations()$Open(<some ppt file>)
slide <- pres$Slides(2)  # the 2nd slide
shape <- slide$Shapes(3)  # the 3rd shape
if (shape[['HasChart']] == -1) {  # if the shape contains msoChart
    chart <- shape[['Chart']]
    refreshPptCharts(NULL, orient="vertical", autoFill=TRUE)
}
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
