% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scale.R, R/set_series.R
\name{set_msochart_scale}
\alias{set_msochart_scale}
\alias{setMsoChartSeries}
\alias{set_msochart_series}
\title{Reset an msoChart object to match the source data and series}
\usage{
set_msochart_scale(oChart, changeScale = TRUE, returnValue = FALSE,
  ...)

setMsoChartSeries(oChart, orient = c("vertical", "horizontal"),
  autoFill = FALSE, ...)

set_msochart_series(oChart, orient = c("vertical", "horizontal"),
  autoFill = FALSE, ...)
}
\arguments{
\item{oChart}{msoChart object}

\item{...}{other arguments, e.g., \code{gapWidth} and \code{emptyString} for
\code{\link{getMsoChartSourceData}}.}

\item{orient}{character, "vertical" or "horizontal". How the series is generated.}

\item{autoFill}{logical, whether fill the series with the entire row/col. Default
FALSE}
}
\value{
Nothing
}
\description{
It fetches meta data from a msoChart object and then reset the series one by
one. The msoChart will be visually modified.
}
\examples{
\dontrun{
ppt <- RDCOMClient::COMCreate("Powerpoint.Application")
pres <- ppt$Presentations()$Open(<some ppt file>)
slide <- pres$Slides(2)  # the 2nd slide
shape <- slide$Shapes(3)  # the 3rd shape
if (shape[['HasChart']] == -1) {  # if the shape contains msoChart
    chart <- shape[['Chart']]
    setMsoChartSeries(chart, orient="vertical", autoFill=TRUE)
}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
